requirements.txt
numpy
pandas
scikit-learn
matplotlib
librosa==0.10.2.post1
audiomentations==0.43.1
pillow
tqdm
pyyaml
soundfile
torch
torchvision
torchaudio
